rpm: support XML database
Some package repositories only have a "primary" link in repodata, not a
"primary_db" with sqlite. In that case, there's an XML file which we
need to parse. That's going to be slower than sqlite, so we should
prefer sqlite wherever possible. But let's implement the "primary" XML
support so that we can support things like Fedora COPR, which don't seem
to have sqlite repository databases.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>