Tweaks to the permalinking implementation
Using Date.getDateString() didn't quite work for me: the dates returned
by the date picker never seemed to show up in the mapping. So use the
simpler YYYY-MM-DD date format and manually format the dates. The result
has been reliable for me.

Another tweak is to store the currentVersion, and once the commits are
loaded, identify the date for this version and update the header to
display a date (more user-friendly than a SHA).

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>