Add undocumented "date" query parameter
Date is nicer to have in the URL: it's more readable, and it actually
has more to do with the information than the version. However, it
requires converting the date into a SHA before we can query the data, so
it's a bit less efficient.

Yet still, I want to support it as an option. So, add the ability to
delay requesting data until the SHAs are loaded. And add the ability to
determine the version from the date on load.

No behavior should really change, unless you go ahead and use the "date"
query parameter rather than version.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>