web: use latest git SHA for date picker
As it is we are overwriting the mapping with the oldest commit on a
particular date. For dates where there are multiple entries, it usually
means we made some updates for a particular new version or distro. As a
result, it's best to prefer the newest commit on that day, not the
oldest. Let's track this and only use the newest date. We can't use the
set of dates (which you'd expect we could) because the JS "Date" object
is actually date + time, and commits are not likely on the exact same
date and time.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>