Switch to pipenv
The requirements-tools project, while nice, has been abandoned by Yelp
for ages. A good approach for applications is pipenv, which can lock
dependencies in the same way requirements-tools. Let's switch to this.
By managing pre-commit in the pipenv development environment, we can
ensure that the hooks run the same locally as they do in CI as well.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>