Settle on Python 3.12 for the "official version"
The README needed updating, and I wanted to provide a way to install
without relying on pipenv. So I added a make target for this and
documented it. But really, pipenv expects that you're just using a
single Python version, and for simplicity, it should be the most recent
version available on OL9. Thus, I've reset all the depnedencies to 3.12.

Now, the README shows a valid way to install (some version of) the
dependencies and get the project running, and using pipenv will set it
up with the exact dependencies too. The CI system will use Python 3.12
too, so we should be set.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>