main: improve --filter operation
For doing local tests while adding distros, I don't really want to do a
_full_ download and extraction of all the distro kernels. I just want to
test the newly added ones. That's what the --filter option is for, but
it is lacking. First, there's no way to specify a pattern, which is
quite useful for specifying a whole set of distros. And second, using
the --filter option prevents state.json from being saved, which means we
lose the opportunity to test out the logic that avoids re-doing work.
Resolve the former by allowing fnmatch patterns, and the latter by
pre-populating the state dictionaries with the values from the old
ones (only when the --filter option is applied).

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>