kconfigs.cleanup: handle non-directory entries
If the analyzer has been previously run, then the output directory
contains a summary.json file. Then if the cleaner runs, there is an
error trying to do an rmtree() on a file.

Since cleanup is designed to remove all unnecessary files, let's just
add support for deleting files like summary.json, and move on with life.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>