kconfigs.{analyzer,cleanup}: support filter logic
One of the annoyances for running kconfigs locally is that the entire
list of distros is quite long. Running it will take a fair bit of
bandwidth and CPU when decompressiong archives, as well as disk space on
a temporary basis. Usually I only really care about testing the distros
I've changed. But only kconfigs.main supports the filter logic.

Add the ability to filter in the other CLIs.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>