config: fedora: add Fedora 43 and clean up
We have had two entries for Fedora "Core" and "Updates", yet I haven't
seen any differences to the kernel in each repo. Even if there is a
difference, I'm not sure it will matter too much, since it'll likely be
due to a new version from a later distro version. To save space, let's
unify everything.

And of course, let's get Fedora 43 included, and drop 40.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>